requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
numpy==1.24.1
outcome==1.2.0
pandas==1.5.2
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
undetected-chromedriver==3.2.1
urllib3==1.26.14
websockets==10.4
wsproto==1.2.0